requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
alembic==1.13.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.8.2
pgvector==0.2.5
httpx==0.27.2
sentence-transformers==3.0.1
torch==2.3.1
pytest==8.3.2
pytest-cov==5.0.0
